examples/fairrrish_project/requirements.txt
numpy==1.26.4
pandas==2.2.1
h5py==3.10.0
matplotlib==3.8.3
pyyaml==6.0.1
